Started working on a packet handler to update client with render & tool tip data, Copied VE's netty code

+++ src/main/java/com/builtbroken/magicmirror/handler/MirrorHandler.java
@@ -143,7 +143,12 @@ public class MirrorHandler
         return userIDToMirrorLocation.get(e.getGameProfile().getId());
     }
 
-    public static int getXpTeleportCost(EntityPlayer player)
+    /**
+     * Gets the cost of teleporting for the user
+     * @param player
+     * @return
+     */
+    public static float getXpTeleportCost(EntityPlayer player)
     {
         if(hasLocation(player))
         {
